bot/client.py
from binance.client import Client
import os
from dotenv import load_dotenv

load_dotenv()

class BinanceClient:

    def __init__(self):
        api_key = os.getenv("BINANCE_API_KEY")
        api_secret = os.getenv("BINANCE_API_SECRET")

        self.client = Client(api_key, api_secret)

        # Testnet URL
        self.client.FUTURES_URL = "https://testnet.binancefuture.com/fapi"