requirements.txt
python-binance
python-dotenv
Flask